Refactor: SignInView Extension 추가
SignInView의 코드 길이가 너무 길어지는 것 같아, SignIn 관련 코드들은 extension에서 관리하도록 했습니다.
